Added parametrized fractal size in benchmark.

+++ run_bench.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+
+if [ -z "$1" ]; then
+    echo "Specify output size!"
+    exit 1
+fi
+
 rm -rf bin/jobs/
-./run.sh "$1" "$2"&
-cd bin && ./cluster_test
+./run.sh "$2" "$3" "$4" "$5"&
+cd bin && ./cluster_test "$1"
